requirements.txt
aioredis==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.1
fastapi-cache==0.1.0
fastapi-cli==0.0.4
file==0.3.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpie==3.2.3
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PySocks==1.7.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid6==2024.7.10
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
